Added PaymentProcessor webjob
Added webjob for procesing payments. This will get data from the
Payments API and get the latest period end values. If there are new
values then a message will be added to a queue for every account
indicating that new payment information should be retrieved. This will
be processed by a seperate worker role