Changed get transactions action to take date range
The transactions details action  now takes a from and to date to get transactions for an account with a given date range.
This has required a change through the entire stack from controller to SQL to support not the submission ID but the date range instead.
Further work is needed to get this working fully as currently the view doesn't supply the  date and the transactions do not store the ranges for the sub-transactions they store.