Updated the Account transactions retrieval code
As part of the refactoring of the account transactions, the repository and levy service code
now supports the retrieval of different types of transactions and now different mediator
calls will retrieve different types of transactions when needed. The transactions details
model has also been replaced with the existing transaction models as they hold the same Data
and the repository now uses a DB model to pull transactions from the database before they are
mapped to the correct transaction type.