English fraction has been added to the PAYE schemes returned in the overview list
The current English fraction for a specific PAYE scheme has been added to the PAYE view model
which gets consumed by the view PAYE Schemes view. This has not been connected to the view yet
but the mediator handler does populate the value. There are a few failing acceptance tests that need to be
fixed around this feature but all unit tests are working.