Initial refactoring of the find organisation section of the website
The new code aims to simplify the logic around different organisation types and how we
manage the find of an organisation process. Previously we tried to make the process as
generic as possible but this made the logic more complex than it needed to be with little
gain so the new code tries to reduce such complexity by treating such organisation differently
where they need to be. The adding of the legal entity is likely broken due to these changes and will
be fixed in a further commit as it also will need to be changed.