Onboarding wire up of new organisation workflow (part complete)
This commit is about 90% functionally working for the new onboarding workflow.
There is plenty of code refactoring to do with duplicate code and some 'faked data'
in the code to make the workflow work. This version should be ok to send to the test team
to at least start the testing of the workflow mechanics but its not ready for Beta release.