Converted all models to view models
As part of the tech debt we looked into how models were being place in the web project.
After discussions it seems the best approach was to rename model into view models so that domain models
are not confused with web view models. The view models are really models in the MVC sense rather than MVVM view models
but have the view model suffix to easily differentiate them from domain models in the code