Add controller action for generating PDF agreement
Added a controller action that will generate the legal agreement based
on a param passed in as to whether the sub fields should be included.
This will be used by a new tool for generating the PDF. This has been
marked as #if DEBUG so that it is not released/available as part of the
release.

Moved the _Agreement_v1 to the shared folder