Con 82 two different apprentices with the same name being aggregated (#1531)
* CON-82 Fix - Using ULN instead of name for grouping

Changed SQL to pull in ULN as well as name.
Changed some data models to allow for ULN to be pulled through to the correct place.
Changed grouping in the orchestrator to group by ULN

* Update EmployerAccountTransactionsOrchestrator.cs

* Revert "Update EmployerAccountTransactionsOrchestrator.cs"

This reverts commit 060ce811dda0dca78f2ed55a324ca465d8089336.

* CON-82 UI Fix to display ULN to prevent confusion

Amended the web view to show the ULN along with apprentice name to prevent confusion created from prior fix. The code in the .cshtml isn't the cleanest but it works front-end.

* Split the null or empty check to run on each value

Split the null or empty to run on both the name and ULN seperately. The ULN can't currently be NULL but no harm in being extra secure

* Revert "Split the null or empty check to run on each value"

This reverts commit 3d1494622fc64b437ee601e296a1b5fd23903d52.

* Revert "CON-82 UI Fix to display ULN to prevent confusion"

This reverts commit 96ddcb5dd0891e541a0807c806c4ffbcdadc09c7.

* Revising UI and code to match demand from UX

Reverting commits and making sure everything is back to the original fix

* [CON-82] Group by ULN only as it is unique